requirements.txt
alabaster==0.7.12
asgiref==3.3.1
astropy==4.0.2
async-generator==1.10
atomicwrites==1.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
bkcharts==0.2
boto==2.49.0
boto3==1.17.27
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cytoolz==0.11.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
glob2==0.7
h5py==2.10.0
HeapDict==1.0.1
heroku3==4.2.3
imagesize==1.2.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
json5==0.9.5
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
mysqlclient==2.0.3
nose==1.3.7
numba==0.51.2
numexpr==2.7.1
olefile==0.46
packaging==20.4
partd==1.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.6.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycosat==0.6.3
pycurl==7.43.0.6
pyflakes==2.2.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtPy==1.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.4.1
statsmodels==0.12.0
tables==3.6.1
terminado==0.9.1
testpath==0.4.4
tifffile==2020.10.1
unicodecsv==0.14.1
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
gunicorn==20.0.4